requirements.txt
requests==2.31.0
pandas==2.2.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pydantic==2.6.3
requests-ratelimiter==0.5.1
PyJWT==2.8.0
PySimpleGUI==5.0.8.3 